+++ source/kibble/random/uuid.cpp
@@ -41,8 +41,8 @@ inline void m128itos(__m128i x, char* mem)
 
     // Add dashes between blocks as specified in RFC-4122
     // 8-4-4-4-12
-    const __m256i dash_shuffle = _mm256_set_epi32(int(0x0b0a0908), int(0x07060504), int(0x80030201), int(0x00808080),
-                                                  int(0x0d0c800b), int(0x0a090880), int(0x07060504), int(0x03020100));
+    const __m256i dash_shuffle = _mm256_set_epi32(0x0b0a0908, 0x07060504, int(0x80030201), 0x00808080, 0x0d0c800b,
+                                                  0x0a090880, 0x07060504, 0x03020100);
     const __m256i dash =
         _mm256_set_epi64x(0x0000000000000000ull, 0x2d000000002d0000ull, 0x00002d000000002d, 0x0000000000000000ull);
 
@@ -64,12 +64,12 @@ inline void m128itos(__m128i x, char* mem)
 inline __m128i stom128i(const char* mem)
 {
     // Remove dashes and pack hex ascii bytes in a 256-bits int
-    const __m256i dash_shuffle = _mm256_set_epi32(int(0x80808080), int(0x0f0e0d0c), int(0x0b0a0908), int(0x06050403),
-                                                  int(0x80800f0e), int(0x0c0b0a09), int(0x07060504), int(0x03020100));
+    const __m256i dash_shuffle = _mm256_set_epi32(int(0x80808080), 0x0f0e0d0c, 0x0b0a0908, 0x06050403, int(0x80800f0e),
+                                                  0x0c0b0a09, 0x07060504, 0x03020100);
 
     __m256i x = betole256(_mm256_loadu_si256(reinterpret_cast<const __m256i*>(mem)));
     x = _mm256_shuffle_epi8(x, dash_shuffle);
-    x = _mm256_insert_epi16(x, betole16(*reinterpret_cast<const uint16_t*>(mem + 16)), 7);
+    x = _mm256_insert_epi16(x, static_cast<short>(betole16(*reinterpret_cast<const uint16_t*>(mem + 16))), 7);
     x = _mm256_insert_epi32(x, betole32(*reinterpret_cast<const uint32_t*>(mem + 32)), 7);
 
     // Build a mask to apply a different offset to alphas and digits
@@ -77,10 +77,10 @@ inline __m128i stom128i(const char* mem)
     const __m256i mask = _mm256_set1_epi8(0x20);
     const __m256i alpha_offset = _mm256_set1_epi8(0x28);
     const __m256i digits_offset = _mm256_set1_epi8(0x01);
-    const __m256i unweave = _mm256_set_epi32(int(0x0f0d0b09), int(0x0e0c0a08), int(0x07050301), int(0x06040200),
-                                             int(0x0f0d0b09), int(0x0e0c0a08), int(0x07050301), int(0x06040200));
-    const __m256i shift = _mm256_set_epi32(int(0x00000000), int(0x00000004), int(0x00000000), int(0x00000004),
-                                           int(0x00000000), int(0x00000004), int(0x00000000), int(0x00000004));
+    const __m256i unweave = _mm256_set_epi32(0x0f0d0b09, 0x0e0c0a08, 0x07050301, 0x06040200, 0x0f0d0b09, 0x0e0c0a08,
+                                             0x07050301, 0x06040200);
+    const __m256i shift = _mm256_set_epi32(0x00000000, 0x00000004, 0x00000000, 0x00000004, 0x00000000, 0x00000004,
+                                           0x00000000, 0x00000004);
 
     // Translate ascii bytes to their value
     // i.e. 0x3132333435363738 -> 0x0102030405060708
